Add namespace and change class name for an cleaner API
Class name change to better fit both read and write functionality and
namespace was added to not conflict with other "json" implementations.

Start to clean upp some C style function name prefix like jRead and swith
to more objectoriented like getElement. But more of this prefix need to
be changed for an cleaner class.